сделал метод Buy который контролирует покупку изменение и отображение денег и в случае удачной покупки вызывает apply
